requirements.txt
# RapidEye - Disaster Detection & Response Prioritization
# Core ML
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.3

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
opencv-python>=4.8.0
albumentations>=1.3.0
scikit-learn>=1.3.0

# Geospatial
rasterio>=1.3.0
geopandas>=0.14.0
shapely>=2.0.0
leafmap>=0.27.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
